+++ lib/src/annotations.dart
@@ -2,4 +2,3 @@
 class FreezedGoStyle {
   const FreezedGoStyle();
 }
-